presenters deleted all done
Presenter_picwall, Presenter_search, Presenter_noteinfo were deleted. All presenters are gone.
