+++ app/src/main/java/com/next/diarycraft/SquareFragment.java
@@ -9,20 +9,27 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.next.diarycraft.Presenter.Presenter_picwall;
+import com.next.diarycraft.Bean.NoteBean;
+import com.next.diarycraft.Model.NoteInfoModel;
+
+import java.util.List;
+
 //图墙fragment photowall fragment
 public class SquareFragment extends Fragment {
     private Toolbar toolbar;
     private AppCompatActivity myActivity;
-    private Presenter_picwall presenter_picwall;
     private MyScrollView myScrollView;
+    private NoteInfoModel noteInfoModel;
+    public String[] ImagePaths;
+    public List<NoteBean> list;
 
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState){
         View view = inflater.inflate(R.layout.fragment_square, container, false);
         toolbar = (Toolbar) view.findViewById (R.id.toolbar_square);
-        presenter_picwall = new Presenter_picwall(this);
+        noteInfoModel=new NoteInfoModel(this.getActivity());
+        refreshPicPaths();
         myScrollView = (MyScrollView) view.findViewById(R.id.my_scroll_view);
-        myScrollView.setPresenter(presenter_picwall,this);
+        myScrollView.setPresenter(this);
         return view;
     }
 
@@ -37,6 +44,15 @@ public class SquareFragment extends Fragment {
     @Override
     public void onResume() {
         super.onResume();
-        presenter_picwall.refreshPicPaths();
+        refreshPicPaths();
+    }
+
+    public void refreshPicPaths(){
+        //获取所有包含图片的notebean和其中所有的图片路径 // get all of the notebeans in the database that contains picture and the path of these pictures.
+        list = noteInfoModel.readAllPhotoBean();
+        ImagePaths = new String[list.size()];
+        for(int i=0;i<list.size();i++){
+            ImagePaths[i] = list.get(i).getPhotopath();
+        }
     }
 }
